Refactor js
